Introduce Analyzers
Analyzers are classes which will take a document and optionally a
position, and return some information about the source we are editing

the AST analyzer uses the vyper compiler to build an AST and provide
data. the source analyzer uses the lark grammar to do the same. they
catch different types of errors, and their callers may decide to ignore
some of their returned results